requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0
nuscenes-devkit>=1.1.0
matplotlib>=3.7.0
numpy>=1.24.0
pyyaml>=6.0
Pillow>=9.5.0
scipy>=1.10.0
tqdm>=4.65.0
pyquaternion
